% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furao.R
\name{.calculate_returns}
\alias{.calculate_returns}
\title{Calculate the returns}
\usage{
.calculate_returns(
  data,
  tickers = "ticker",
  dates = "ref_date",
  prices = "price_adjusted",
  returns = "return"
)
}
\arguments{
\item{data}{Data frame in long format.}

\item{tickers}{Name of the column which holds tickers names. (default = "ticker")}

\item{dates}{Name of the column which holds dates. (default = "ref_date")}

\item{prices}{Name of the column which holds stock prices. (default = "price_adjusted")}

\item{returns}{Name of the column in which calculated returns will be stored. (default = "return")}
}
\description{
Calculate the returns
}
