% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furao.R
\name{.set_frequency}
\alias{.set_frequency}
\title{Delete rows so the remaining observations follow a determined time interval}
\usage{
.set_frequency(data, interval = "month", dates = "ref_date", slice_by = c())
}
\arguments{
\item{data}{Data frame in long format.}

\item{interval}{Frequency of data: "daily", "weekly", "monthly", "yearly". (default = "month")}

\item{dates}{Name of the column which holds dates. (default = "ref_date")}

\item{slice_by}{Vector of columns to group by for just this operation.}
}
\description{
Delete rows so the remaining observations follow a determined time interval
}
