% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furao.R
\name{.sharpe_beta}
\alias{.sharpe_beta}
\title{Calculate Sharpe's Beta for a given portfolio relative to a market}
\usage{
.sharpe_beta(data, portfolio_returns, market_returns)
}
\arguments{
\item{data}{Data frame in wide format.}

\item{portfolio_returns}{Name of the column which holds the portfolio returns.}

\item{market_returns}{Name of the column which holds the market returns.}
}
\description{
Calculate Sharpe's Beta for a given portfolio relative to a market
}
