% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furao.R
\name{get_rolling_betas}
\alias{get_rolling_betas}
\title{Calculate Rolling Betas}
\usage{
get_rolling_betas(
  data,
  .dates_col = "ref_date",
  .prices_col = "price_adjusted",
  .tickers_col = "ticker",
  .portfolio_tickers,
  .weights = NULL,
  .market_ticker,
  .beta_fun = .sharpe_beta,
  .observations = 36L,
  .interval = "month",
  .date_lb = NULL,
  .date_ub = NULL,
  .beta_adj = NULL,
  .diagnostics = FALSE
)
}
\arguments{
\item{data}{Data frame in long format.}

\item{.dates_col}{Name of the column which holds the dates.}

\item{.prices_col}{Name of the column which holds the tickers prices.}

\item{.tickers_col}{Name of the column which holds the tickers names.}

\item{.portfolio_tickers}{String or vector specifying the ticker(s) the Beta is being calculated for.}

\item{.weights}{Vector of integers specifying the weights for each ticker in the portfolio. It's internally normalized so the weights sums to 1. Defaults to 1.}

\item{.market_ticker}{Name of the ticker which the portfolio is being regressed against.}

\item{.beta_fun}{Function definition of the Beta. (default = Sharpe's Beta)}

\item{.observations}{Number of observations used to calculate each Beta. (default = 36L)}

\item{.interval}{Desired frequency of returns: "daily", "weekly", "monthly", "yearly". (default = "month")}

\item{.date_lb}{Begin point for which the data.frame is truncated for. (inclusive. If left empty will use the entire dataset)}

\item{.date_ub}{End point for which the data.frame is truncated for. (inclusive. If left empty will use the entire dataset)}

\item{.beta_adj}{Adjust method for Beta. (not implemented yet, will raise an error if filled.)}

\item{.diagnostics}{Returns plots and test statistics.}
}
\value{
A list
}
\description{
Calculate Rolling Betas
}
